api/requirements.txt
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
firebase-admin==5.0.3
Flask==2.0.2
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
httplib2==0.20.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
proto-plus==1.19.5
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
psycopg2-binary==2.9.1
populartimes @ git+https://github.com/m-wrzr/populartimes@3d3df3f306b6f65d7cf9f5c9387ea665bdfbf836
sqlalchemy
sqlalchemy-cockroachdb==1.4.2
geopy==2.2.0